CameraPose结构体内的R与t修改为CV_64F

+++ src/CPU/SFMTypes.h
@@ -5,8 +5,8 @@
 
 struct CameraPose 
 {
-    cv::Mat R = cv::Mat::eye(3, 3, CV_32F);
-    cv::Mat t = cv::Mat::zeros(3, 1, CV_32F);
+    cv::Mat R = cv::Mat::eye(3, 3, CV_64F);
+    cv::Mat t = cv::Mat::zeros(3, 1, CV_64F);
 };
 
 struct SFMResult
